refactor: change structure of map folder component
