+++ src/components/map/Map.css
@@ -1,6 +1,7 @@
 .map-container {
-    width: 100%;
-    height: 100vh;
+    width: 800px;
+    height: 800px;
+    border-radius: 10px;
   }
 
   .map {
