+++ src/components/map/mapbox/index.js
@@ -1,0 +1,3 @@
+import MapBox from './MapBox';
+
+export default MapBox;
