+++ src/components/map/mapbox/popup/Popup.js
@@ -1,9 +1,10 @@
+import './Popup.css';
 
 const Popup = ({ feature }) => {
     const { id, name, description, phone, address, address2, address3 } = feature.properties;
 
     return (
-        <div id={`popup-${id}`}>
+        <div id={`popup-${id}`} className="map-pop-up">
             <h3>{name}</h3>
             <p>{description}</p>
             <p>{phone}</p>
